+++ src/test/java/com/jobnote/auth/LoginFilterApplicationTest.java
@@ -1,4 +1,4 @@
-package com.jobnote.auth.security;
+package com.jobnote.auth;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jobnote.JobnoteApplicationTests;
